Updated README.md
I changed step 6 and 7 to be more explicit about exactly where to find the appropriate release page. I'm sure the appropriate action is obvious to the more experienced but I have never touched the release page before and did not understand how that part of git hub's structure worked. I hope my syntax is ok, if not I'd be very interested in seeing a more appropriate way of implementing my recommendation.